requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-huggingface
faiss-cpu
pypdf
pymupdf
tabula-py
pytesseract
pdf2image
opencv-python
pandas
numpy
python-dotenv
langchain_community
tabulate
jpype1
sentence-transformers